backend/requirements.txt
Django==5.2.1
# Use a DRF patch release to satisfy other packages that require >=3.15.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.5.1
django-filter==24.3
django-ckeditor-5==0.1.1
django-cors-headers==4.0.0
django-debug-toolbar==5.2
# Pillow bumped to satisfy pillow-heif and to avoid building older wheels
Pillow==11.3.0
requests==2.32.5
